benchmark: refactor some code to avoid duplications
