+++ src/uu/numfmt/benches/numfmt_bench.rs
@@ -4,45 +4,32 @@
 // file that was distributed with this source code.
 
 use divan::{Bencher, black_box};
-use tempfile::TempDir;
 use uu_numfmt::uumain;
-use uucore::benchmark::{create_test_file, run_util_function};
-
-/// Generate numeric data for benchmarking
-fn generate_numbers(count: usize) -> String {
-    (1..=count)
-        .map(|n| n.to_string())
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
-/// Setup benchmark environment with test data
-fn setup_benchmark(data: String) -> (TempDir, String) {
-    let temp_dir = tempfile::tempdir().unwrap();
-    let file_path = create_test_file(data.as_bytes(), temp_dir.path());
-    let file_path_str = file_path.to_str().unwrap().to_string();
-    (temp_dir, file_path_str)
-}
+use uucore::benchmark::{run_util_function, setup_test_file, text_data};
 
 /// Benchmark SI formatting with different number counts
 #[divan::bench(args = [1_000_000])]
 fn numfmt_to_si(bencher: Bencher, count: usize) {
-    let (_temp_dir, file_path_str) = setup_benchmark(generate_numbers(count));
+    let data = text_data::generate_numbers(count);
+    let file_path = setup_test_file(data.as_bytes());
+    let file_path_str = file_path.to_str().unwrap();
 
     bencher.bench(|| {
-        black_box(run_util_function(uumain, &["--to=si", &file_path_str]));
+        black_box(run_util_function(uumain, &["--to=si", file_path_str]));
     });
 }
 
 /// Benchmark SI formatting with precision format
 #[divan::bench(args = [1_000_000])]
 fn numfmt_to_si_precision(bencher: Bencher, count: usize) {
-    let (_temp_dir, file_path_str) = setup_benchmark(generate_numbers(count));
+    let data = text_data::generate_numbers(count);
+    let file_path = setup_test_file(data.as_bytes());
+    let file_path_str = file_path.to_str().unwrap();
 
     bencher.bench(|| {
         black_box(run_util_function(
             uumain,
-            &["--to=si", "--format=%.6f", &file_path_str],
+            &["--to=si", "--format=%.6f", file_path_str],
         ));
     });
 }
@@ -50,10 +37,12 @@ fn numfmt_to_si_precision(bencher: Bencher, count: usize) {
 /// Benchmark IEC (binary) formatting
 #[divan::bench(args = [1_000_000])]
 fn numfmt_to_iec(bencher: Bencher, count: usize) {
-    let (_temp_dir, file_path_str) = setup_benchmark(generate_numbers(count));
+    let data = text_data::generate_numbers(count);
+    let file_path = setup_test_file(data.as_bytes());
+    let file_path_str = file_path.to_str().unwrap();
 
     bencher.bench(|| {
-        black_box(run_util_function(uumain, &["--to=iec", &file_path_str]));
+        black_box(run_util_function(uumain, &["--to=iec", file_path_str]));
     });
 }
 
@@ -65,10 +54,11 @@ fn numfmt_from_si(bencher: Bencher, count: usize) {
         .map(|n| format!("{:.1}K", n as f64 / 1000.0))
         .collect::<Vec<_>>()
         .join("\n");
-    let (_temp_dir, file_path_str) = setup_benchmark(data);
+    let file_path = setup_test_file(data.as_bytes());
+    let file_path_str = file_path.to_str().unwrap();
 
     bencher.bench(|| {
-        black_box(run_util_function(uumain, &["--from=si", &file_path_str]));
+        black_box(run_util_function(uumain, &["--from=si", file_path_str]));
     });
 }
 
@@ -80,23 +70,26 @@ fn numfmt_large_numbers_si(bencher: Bencher, count: usize) {
         .map(|n| (n * 1_000_000).to_string())
         .collect::<Vec<_>>()
         .join("\n");
-    let (_temp_dir, file_path_str) = setup_benchmark(data);
+    let file_path = setup_test_file(data.as_bytes());
+    let file_path_str = file_path.to_str().unwrap();
 
     bencher.bench(|| {
-        black_box(run_util_function(uumain, &["--to=si", &file_path_str]));
+        black_box(run_util_function(uumain, &["--to=si", file_path_str]));
     });
 }
 
 /// Benchmark different padding widths
 #[divan::bench(args = [(1_000_000, 5), (1_000_000, 50)])]
 fn numfmt_padding(bencher: Bencher, (count, padding): (usize, usize)) {
-    let (_temp_dir, file_path_str) = setup_benchmark(generate_numbers(count));
+    let data = text_data::generate_numbers(count);
+    let file_path = setup_test_file(data.as_bytes());
+    let file_path_str = file_path.to_str().unwrap();
     let padding_arg = format!("--padding={padding}");
 
     bencher.bench(|| {
         black_box(run_util_function(
             uumain,
-            &["--to=si", &padding_arg, &file_path_str],
+            &["--to=si", &padding_arg, file_path_str],
         ));
     });
 }
@@ -104,13 +97,15 @@ fn numfmt_padding(bencher: Bencher, (count, padding): (usize, usize)) {
 /// Benchmark round modes with SI formatting
 #[divan::bench(args = [("up", 100_000), ("down", 1_000_000), ("towards-zero", 1_000_000)])]
 fn numfmt_round_modes(bencher: Bencher, (round_mode, count): (&str, usize)) {
-    let (_temp_dir, file_path_str) = setup_benchmark(generate_numbers(count));
+    let data = text_data::generate_numbers(count);
+    let file_path = setup_test_file(data.as_bytes());
+    let file_path_str = file_path.to_str().unwrap();
     let round_arg = format!("--round={round_mode}");
 
     bencher.bench(|| {
         black_box(run_util_function(
             uumain,
-            &["--to=si", &round_arg, &file_path_str],
+            &["--to=si", &round_arg, file_path_str],
         ));
     });
 }
