+++ src/uu/sort/benches/sort_locale_bench.rs
@@ -5,79 +5,14 @@
 
 use divan::{Bencher, black_box};
 use std::env;
-use std::path::PathBuf;
 use uu_sort::uumain;
-use uucore::benchmark::{create_test_file, run_util_function};
-
-/// Helper function to generate test data from a list of words
-fn generate_data_from_words_with_counter(words: &[&str], num_lines: usize) -> Vec<u8> {
-    let mut data = Vec::new();
-    for i in 0..num_lines {
-        let word = words[i % words.len()];
-        let line = format!("{word}{i:04}\n");
-        data.extend_from_slice(line.as_bytes());
-    }
-    data
-}
-
-fn generate_ascii_data(num_lines: usize) -> Vec<u8> {
-    let mut data = Vec::new();
-    for i in 0..num_lines {
-        let line = format!("line_{:06}\n", (num_lines - i - 1));
-        data.extend_from_slice(line.as_bytes());
-    }
-    data
-}
-
-fn generate_mixed_locale_data(num_lines: usize) -> Vec<u8> {
-    let mixed_strings = [
-        "zebra", "äpfel", "banana", "öl", "cat", "über", "dog", "zürich", "elephant", "café",
-        "fish", "naïve", "grape", "résumé", "house", "piñata",
-    ];
-    generate_data_from_words_with_counter(&mixed_strings, num_lines)
-}
-
-fn generate_german_locale_data(num_lines: usize) -> Vec<u8> {
-    let german_words = [
-        "Ärger", "Öffnung", "Über", "Zucker", "Bär", "Föhn", "Größe", "Höhe", "Käse", "Löwe",
-        "Mädchen", "Nüsse", "Röntgen", "Schäfer", "Tür", "Würfel", "ä", "ö", "ü", "ß", "a", "o",
-        "u", "s",
-    ];
-    generate_data_from_words_with_counter(&german_words, num_lines)
-}
-
-fn generate_random_strings(num_lines: usize, length: usize) -> Vec<u8> {
-    let mut data = Vec::new();
-    let charset =
-        "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789äöüÄÖÜßéèêëàâîïôûç";
-    let charset_bytes = charset.as_bytes();
-
-    for i in 0..num_lines {
-        let mut line = String::new();
-        for j in 0..length {
-            let idx = ((i * length + j) * 17 + 42) % charset_bytes.len();
-            line.push(charset_bytes[idx] as char);
-        }
-        line.push('\n');
-        data.extend_from_slice(line.as_bytes());
-    }
-    data
-}
-
-fn setup_test_file(data: &[u8]) -> PathBuf {
-    let temp_dir = tempfile::tempdir().unwrap();
-    let file_path = create_test_file(data, temp_dir.path());
-    // Keep temp_dir alive by leaking it - the OS will clean it up
-    std::mem::forget(temp_dir);
-    file_path
-}
+use uucore::benchmark::{run_util_function, setup_test_file, text_data};
 
 /// Benchmark ASCII-only data sorting with C locale (byte comparison)
 #[divan::bench]
 fn sort_ascii_c_locale(bencher: Bencher) {
-    let temp_dir = tempfile::tempdir().unwrap();
-    let data = generate_ascii_data(100_000);
-    let file_path = create_test_file(&data, temp_dir.path());
+    let data = text_data::generate_ascii_data_simple(100_000);
+    let file_path = setup_test_file(&data);
 
     bencher.bench(|| {
         unsafe {
@@ -90,7 +25,7 @@ fn sort_ascii_c_locale(bencher: Bencher) {
 /// Benchmark ASCII-only data sorting with UTF-8 locale
 #[divan::bench]
 fn sort_ascii_utf8_locale(bencher: Bencher) {
-    let data = generate_ascii_data(10_000);
+    let data = text_data::generate_ascii_data_simple(10_000);
     let file_path = setup_test_file(&data);
 
     bencher.bench(|| {
@@ -104,7 +39,7 @@ fn sort_ascii_utf8_locale(bencher: Bencher) {
 /// Benchmark mixed ASCII/Unicode data with C locale
 #[divan::bench]
 fn sort_mixed_c_locale(bencher: Bencher) {
-    let data = generate_mixed_locale_data(10_000);
+    let data = text_data::generate_mixed_locale_data(10_000);
     let file_path = setup_test_file(&data);
 
     bencher.bench(|| {
@@ -118,7 +53,7 @@ fn sort_mixed_c_locale(bencher: Bencher) {
 /// Benchmark mixed ASCII/Unicode data with UTF-8 locale
 #[divan::bench]
 fn sort_mixed_utf8_locale(bencher: Bencher) {
-    let data = generate_mixed_locale_data(10_000);
+    let data = text_data::generate_mixed_locale_data(10_000);
     let file_path = setup_test_file(&data);
 
     bencher.bench(|| {
@@ -132,7 +67,7 @@ fn sort_mixed_utf8_locale(bencher: Bencher) {
 /// Benchmark German locale-specific data with C locale
 #[divan::bench]
 fn sort_german_c_locale(bencher: Bencher) {
-    let data = generate_german_locale_data(10_000);
+    let data = text_data::generate_german_locale_data(10_000);
     let file_path = setup_test_file(&data);
 
     bencher.bench(|| {
@@ -146,7 +81,7 @@ fn sort_german_c_locale(bencher: Bencher) {
 /// Benchmark German locale-specific data with German locale
 #[divan::bench]
 fn sort_german_locale(bencher: Bencher) {
-    let data = generate_german_locale_data(10_000);
+    let data = text_data::generate_german_locale_data(10_000);
     let file_path = setup_test_file(&data);
 
     bencher.bench(|| {
@@ -160,7 +95,7 @@ fn sort_german_locale(bencher: Bencher) {
 /// Benchmark random strings of different lengths
 #[divan::bench]
 fn sort_random_strings(bencher: Bencher) {
-    let data = generate_random_strings(10_000, 50);
+    let data = text_data::generate_random_strings(10_000, 50);
     let file_path = setup_test_file(&data);
 
     bencher.bench(|| {
@@ -195,7 +130,7 @@ fn sort_numeric(bencher: Bencher) {
 /// Benchmark reverse sorting
 #[divan::bench]
 fn sort_reverse_mixed(bencher: Bencher) {
-    let data = generate_mixed_locale_data(10_000);
+    let data = text_data::generate_mixed_locale_data(10_000);
     let file_path = setup_test_file(&data);
 
     bencher.bench(|| {
@@ -212,7 +147,7 @@ fn sort_reverse_mixed(bencher: Bencher) {
 /// Benchmark unique sorting
 #[divan::bench]
 fn sort_unique_mixed(bencher: Bencher) {
-    let data = generate_mixed_locale_data(10_000);
+    let data = text_data::generate_mixed_locale_data(10_000);
     let file_path = setup_test_file(&data);
 
     bencher.bench(|| {
